requirements.txt
numpy>=1.24
opencv-python-headless==4.10.0.84

# Mediapipe + protobuf pin to avoid MessageFactory/GetPrototype crash
mediapipe==0.10.9
protobuf>=3.20.0,<3.21.0

# CUDA ASR via CTranslate2 (no Torch needed)
faster-whisper>=1.0.0
ctranslate2>=3.24.0
sentencepiece>=0.1.99

# Google Drive
google-api-python-client>=2.129.0
google-auth>=2.35.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1

tqdm>=4.66.5
